xml_parser_function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.28.0
pytz==2018.7
google-cloud-error-reporting==0.33.0
google-cloud-pubsub==1.4.2
untangle
google-cloud-logging
pymysql
